chore(de): Consistent capitalization when addressing the user
In German, there are several options of capitalization when addressing the user. The informal form "du" *can* be capitalized, but this is unusual (it's sometimes used when writing a letter, see https://www.duden.de/rechtschreibung/du). Since the German translations contained cases of both upper and lower-case salutations, I revised the file and changed all cases of "du", "dir", "dich" and "dein" to the lower case form. This might be picky or the upper case form might have been chosen by design, that's why I put these changes in a separate commit.